build fix
svn:sync-xref-src-repo-rev: r59225

+++ src/VBox/Runtime/common/string/strformatrt.cpp
@@ -1,4 +1,4 @@
-/* $Id: strformatrt.cpp 27648 2010-03-23 22:42:29Z knut.osmundsen@oracle.com $ */
+/* $Id: strformatrt.cpp 27650 2010-03-23 23:07:08Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - IPRT String Formatter Extensions.
  */
@@ -1053,7 +1053,7 @@ size_t rtstrFormatRt(PFNRTSTROUTPUT pfnOutput, void *pvArgOutput, const char **p
                         static const char   s_szElemEscape[] = "<>&'\"";
                         const char *pszEscape = fAttr ?             s_szAttrEscape      :             s_szElemEscape;
                         int         cchEscape = fAttr ? RT_ELEMENTS(s_szAttrEscape) - 1 : RT_ELEMENTS(s_szElemEscape) - 1;
-                        int         cchOutput = 0;
+                        size_t      cchOutput = 0;
                         const char *pszStr    = va_arg(*pArgs, char *);
                         int         cchStr;
                         int         offCur;
